+++ tyfbaf/__init__.py
@@ -7,4 +7,10 @@ __version__ = "0.0.6"
 
 
 def setup(server_name: str, *, port: int = 6405):
+    """Function to setup tyfbaf.
+
+    Args:
+        server_name (str): Just your server name, without any protocol or port.
+        port (int, optional): The port in use of your server. Defaults to 6405.
+    """
     _setup(server_name, port)